% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{as_proj_area}
\alias{as_proj_area}
\title{Summarise a geographic area}
\usage{
as_proj_area(handleable)
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[wk:wkb]{wkb()}}, \code{\link[wk:wkt]{wkt()}}, \code{\link[wk:xy]{xy()}},
\code{\link[wk:rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[wk:wk_handle]{wk_handle()}} is defined.}
}
\value{
\code{c(west, south, east, north)} in WGS84
}
\description{
The concept of an "area" in PROJ means a geographic one
that can span the antimeridian. This function uses some
heuristics to find the minimum longitude span of the
input (provided that the longitude span is less than
180 degrees). The \code{\link[wk:wk_crs]{wk::wk_crs()}} of the input is
considered and the appropriate transform to WGS84
is applied.
}
\examples{
as_proj_area(wk::rct(-1, -1, 1, 1))
as_proj_area(wk::rct(-179, 179, 1, 1))

}
