% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-wk.R
\name{as_wk_trans.rlibproj_proj}
\alias{as_wk_trans.rlibproj_proj}
\alias{wk_trans_inverse.rlibproj_trans_proj}
\title{Get wk package transforms}
\usage{
\method{as_wk_trans}{rlibproj_proj}(x, ..., use_z = NA, use_m = NA)

\method{wk_trans_inverse}{rlibproj_trans_proj}(trans)
}
\arguments{
\item{x}{A \link[=proj_create]{PROJ object}}

\item{...}{Unused}

\item{use_z, use_m}{Use \code{TRUE} to force inclusion of this dimension
in the result or \code{FALSE} to explicitly omit it. Use \code{NA}
to leave the dimensions of the output the same as the input.
Use a length two vector to specify the dimensionality
of the inverse transform.}

\item{trans}{A transform created with \code{\link[wk:wk_trans_inverse]{wk::as_wk_trans()}}
on a \link[=proj_create]{PROJ object}.}
}
\value{
A \code{\link[wk:wk_trans_inverse]{wk::new_wk_trans()}}
}
\description{
Get wk package transforms
}
