% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-crs2crs.R
\name{crs_engine_rproj}
\alias{crs_engine_rproj}
\alias{crs_engine_proj_pipeline.rlibrpoj_crs2crs_engine}
\alias{crs_engine_proj_pipeline_apply.rlibrpoj_crs2crs_engine}
\alias{crs_engine_get_wk_trans.rlibrpoj_crs2crs_engine}
\alias{crs_proj_definition.rproj_proj}
\title{crs2crs transform engine}
\usage{
crs_engine_rproj(ctx = proj_context(), spatial_test = "intersects")

\method{crs_engine_proj_pipeline}{rlibrpoj_crs2crs_engine}(
  engine,
  handleable,
  crs_to,
  crs_from = wk::wk_crs(handleable),
  bbox = wk::wk_bbox(handleable),
  accuracy = NA_real_,
  auth_name = NA_character_,
  allow_ballpark = NA,
  ...
)

\method{crs_engine_proj_pipeline_apply}{rlibrpoj_crs2crs_engine}(engine, handleable, pipeline, ...)

\method{crs_engine_get_wk_trans}{rlibrpoj_crs2crs_engine}(engine, handleable, crs_to, crs_from, ...)

\method{crs_proj_definition}{rproj_proj}(crs, proj_version = NULL)
}
\arguments{
\item{ctx}{A \code{\link[=proj_context]{proj_context()}}}

\item{spatial_test}{Use "none" to skip using extents when selecting
the appropriate transform between crs objects.}

\item{engine}{A transform engine such as \code{\link[crs2crs:crs_engine_null]{crs_engine_null()}}}

\item{handleable}{A geometry vector (e.g., \code{\link[wk:wkb]{wkb()}}, \code{\link[wk:wkt]{wkt()}}, \code{\link[wk:xy]{xy()}},
\code{\link[wk:rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[wk:wk_handle]{wk_handle()}} is defined.}

\item{crs_to}{Source and destination coordinate reference systems}

\item{crs_from}{Source and destination coordinate reference systems}

\item{bbox}{The bounding box to use when searching for appropriate
transformations.}

\item{accuracy}{The minimum desired accuracy for the transformation}

\item{auth_name}{An authority name (e.g., "EPSG" or "OGC")}

\item{allow_ballpark}{Use \code{FALSE} to omit ballpark transformations}

\item{...}{Unused}

\item{pipeline}{A character vector representing a PROJ coordinate operation}

\item{crs}{An arbitrary R object}

\item{proj_version}{A \code{\link[=package_version]{package_version()}} of the PROJ version, or
\code{NULL} if the PROJ version is unknown.}
}
\value{
An engine that can be used with \code{\link[=crs_set_engine]{crs_set_engine()}}.
}
\description{
crs2crs transform engine
}
