% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{noaa df}
}
\description{
Ha ha ha
}
\details{
Ne ne ne
}
\examples{
\dontrun{eq_create_label(noaa_df)}
}
