% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{noaa df}
}
\description{
Ha ha ha
}
\details{
Ne ne ne
}
\examples{
\dontrun{eq_location_clean(noaa_df)}
}
