% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{eq_plot_map}
\alias{eq_plot_map}
\title{eq_plot_ex_map}
\usage{
eq_plot_map(countries = c("MEXICO", "GUATEMALA"), lbyear = 2000)
}
\arguments{
\item{countries}{list of countries}

\item{lbyear}{Lower bound - years}
}
\description{
eq_plot_ex_map
}
