% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{eq_plot_timeline}
\alias{eq_plot_timeline}
\title{eq_plot_timeline}
\usage{
eq_plot_timeline(countries = c("MEXICO", "GUATEMALA"), lbyear = 2000)
}
\arguments{
\item{countries}{list of countries}

\item{lbyear}{Lower bound - years}
}
\description{
Ha ha ha
}
\details{
Ne ne ne
}
\examples{
\dontrun{eq_plot_timeline()}

}
