% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{eq_plot_timeline_w_labels}
\alias{eq_plot_timeline_w_labels}
\title{eq_plot_timeline_w_labels}
\usage{
eq_plot_timeline_w_labels(countries = c("MEXICO", "GUATEMALA"),
  lbyear = 2000)
}
\arguments{
\item{countries}{list of countries}

\item{lbyear}{Lower bound - years}
}
\description{
eq_plot_timeline_w_labels
}
