% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label function definition}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, n_max = 5, ...)
}
\description{
Using new geom class we can plot the timeline labels
}
\examples{
\dontrun{
 ggplot(noaa_df, aes(x = DATE, y = COUNTRY,
        color = as.numeric(TOTAL_DEATHS),
        size = as.numeric(EQ_PRIMARY),
        label = LOCATION_NAME)) +
 geom_timeline_label()
}
}
