% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline class definition}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeline
}
\description{
ggproto() constructs a new geom class
}
\details{
Ne ne ne
}
\examples{
\dontrun{
 No example, this is part of ggplot2 geom function defintion
}
}
\keyword{datasets}
