% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{GeomTimelineLabel class definition}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimelineLabel
}
\description{
ggproto() constructs a new geom class
}
\examples{
\dontrun{
 No example, this is part of ggplot2 geom function defintion
}
}
\keyword{datasets}
