% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_course.R
\name{install_from_rproj}
\alias{install_from_rproj}
\title{Install a course from the official course repository}
\usage{
install_from_rproj(course_name, dev = FALSE, mirror = "github")
}
\arguments{
\item{course_name}{The name of the course you wish to install.}

\item{dev}{Set to \code{TRUE} to install a course in development from the rproj_misc repository.}

\item{mirror}{Select rproj course repository mirror. Valid arguments are \code{"github"} and \code{"bitbucket"}.}
}
\description{
We are currently maintaining a central repository of contributed
rproj courses at \url{https://github.com/rprojdev/rproj_courses}.
This function provides the easiest method of installing a course
form the repository.

We have another repository at 
\url{https://github.com/rprojdev/rproj_misc}, where we keep 
experimental features and content. The \code{dev} argument allows
you to access this repository. Content in the rproj_misc repository
is not guaranteed to work.

The central repository of rproj courses is mirrored at
\url{https://bitbucket.org/rprojdevmirror/rproj_courses}. If you cannot 
access GitHub you can download rproj courses from bitbucket by using the
\code{mirror = "bitbucket"} option (see below).
}
\examples{
\dontrun{

install_from_rproj("R_Programming") # Directory name

### OR ###

install_from_rproj("R Programming") # Course name

# To install a course in development from the rproj_misc repository
install_from_rproj("Including Data", dev = TRUE)

# To install a course from the Bitbucket mirror
install_from_rproj("R Programming", mirror = "bitbucket")
}
}
\seealso{
Other InstallCourses: 
\code{\link{InstallCourses}},
\code{\link{install_course_directory}()},
\code{\link{install_course_dropbox}()},
\code{\link{install_course_github}()},
\code{\link{install_course_google_drive}()},
\code{\link{install_course_url}()},
\code{\link{install_course_zip}()},
\code{\link{install_course}()},
\code{\link{uninstall_all_courses}()},
\code{\link{uninstall_course}()},
\code{\link{zip_course}()}
}
\concept{InstallCourses}
