% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{proj_coord}
\alias{proj_coord}
\title{Create coordinate matrices}
\usage{
proj_coord(x = NA_real_, y = NA_real_, z = NA_real_, t = NA_real_)
}
\arguments{
\item{x, y, z, t}{Coordinate values}
}
\value{
A matrix with four columns (x, y, z, t)
}
\description{
Create coordinate matrices
}
\examples{
proj_coord(1, 2)

}
