% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_create}
\alias{proj_create}
\alias{proj_create_crs_to_crs}
\alias{proj_create_from_wkt}
\alias{proj_create_from_name}
\alias{proj_guess_wkt_dialect}
\alias{as_proj}
\alias{as_proj.rlibproj_proj}
\alias{as_proj.character}
\title{Create PROJ objects}
\usage{
proj_create(definition, ctx = proj_context())

proj_create_crs_to_crs(
  source_crs,
  target_crs,
  area = NULL,
  auth_name = NA_character_,
  allow_ballpark = NA,
  accuracy = NA_real_,
  ctx = proj_context()
)

proj_create_from_wkt(wkt, ctx = proj_context())

proj_create_from_name(
  searched_name,
  auth_name,
  types,
  approximate_match,
  limit_result_count,
  options = NULL,
  ctx = proj_context()
)

proj_guess_wkt_dialect(wkt, ctx = proj_context())

as_proj(x, ..., ctx = proj_context())

\method{as_proj}{rlibproj_proj}(x, ..., ctx = proj_context())

\method{as_proj}{character}(x, ..., ctx = proj_context())
}
\arguments{
\item{definition}{A character vector definition. This can be
a PROJ string, WKT, Authority:Code, or any other format PROJ
understands.}

\item{ctx}{A \code{\link[=proj_context]{proj_context()}}}

\item{source_crs, target_crs}{Source and/or target CRS definitions,
coerced using \code{\link[=as_proj]{as_proj()}}.}

\item{area}{A \link[=as_proj_area]{PROJ area} to use when selecting the
best available transformation.}

\item{auth_name}{An authority name (e.g., "EPSG" or "OGC")}

\item{allow_ballpark}{Use \code{FALSE} to omit ballpark transformations}

\item{accuracy}{The minimum desired accuracy for the transformation}

\item{wkt}{A well-known text definition of the PROJ object.}

\item{searched_name}{A name query}

\item{types}{The types of objects that should be queried. Type
values of "crs" and "ellipsoid" are probably the most useful.}

\item{approximate_match}{Use \code{TRUE} for near matches}

\item{limit_result_count}{A limit on the number of matches}

\item{options}{Options for instantiating the object by name}

\item{x}{An object to convert to a PROJ pointer}

\item{...}{Passed to S3 methods}
}
\description{
In PROJ speak, a "proj" can be a transformation, conversion,
CRS, ellipsoid, datum, or one of a few other types. The
CRS and conversion types are the most common and the
most useful but the other types are also occasionally
returned.
}
