% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{proj_trans}
\alias{proj_trans}
\title{Transform coordinates}
\usage{
proj_trans(pj, coord, direction = "fwd", verbose = FALSE)
}
\arguments{
\item{pj}{A PROJ object or definition coerced using \code{\link[=as_proj]{as_proj()}}.}

\item{direction}{One of "fwd", "ident", or "inv".}

\item{verbose}{Use \code{FALSE} to keep log output during transforms.
This is usually not what you want, since you might be transforming
millions of coordinates (which may result in millions of log
messages).}

\item{x}{An R object}

\item{...}{Passed to S3 methods}
}
\value{
A transformed value of \code{x}
}
\description{
Transform coordinates
}
\examples{
p <- proj_create_crs_to_crs("OGC:CRS84", "EPSG:3857")
x <- proj_coord(-64, 45)
proj_trans(p, x)

}
