% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rproj-package.R
\docType{package}
\name{rproj-package}
\alias{rproj}
\alias{rproj-package}
\title{rproj: Generic Coordinate Transformation Library ('PROJ') R API}
\description{
Provides R bindings for the 'PROJ' coordinate transformation library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/rproj}
  \item Report bugs at \url{https://github.com/paleolimbot/rproj/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
