% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rproj_options}
\alias{rproj_options}
\title{Get rproj options}
\usage{
rproj_options(...)
}
\arguments{
\item{...}{any options can be defined, using name = value.}
}
\description{
This function is a wrapper for \code{options()} that allows the user to
see the state of how certain options for rproj are set up.
}
\examples{
\dontrun{
# See current current rproj options
rproj_options()

# Set an option
rproj_options(rproj_logging = TRUE)
}
}
