% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_create}
\alias{proj_create}
\alias{proj_create_crs_to_crs}
\alias{proj_create_from_wkt}
\alias{proj_guess_wkt_dialect}
\alias{proj_get_source_crs}
\alias{proj_get_target_crs}
\alias{proj_identify}
\alias{proj_get_non_deprecated}
\alias{proj_normalize_for_visualization}
\alias{as_proj}
\alias{as_proj.rlibproj_proj}
\alias{as_proj.character}
\title{Create PROJ objects}
\usage{
proj_create(definition, ctx = proj_context())

proj_create_crs_to_crs(
  source_crs,
  target_crs,
  area = NULL,
  options = character(),
  ctx = proj_context()
)

proj_create_from_wkt(wkt, options = character(), ctx = proj_context())

proj_guess_wkt_dialect(wkt, ctx = proj_context())

proj_get_source_crs(pj)

proj_get_target_crs(pj)

proj_identify(pj, auth_name, options = NULL, ctx = proj_context())

proj_get_non_deprecated(pj, ctx = proj_context())

proj_normalize_for_visualization(pj, ctx = proj_context())

as_proj(x, ..., ctx = proj_context())

\method{as_proj}{rlibproj_proj}(x, ..., ctx = proj_context())

\method{as_proj}{character}(x, ..., ctx = proj_context())
}
\arguments{
\item{definition}{A character vector definition. This can be
a PROJ string, WKT, Authority:Code, or any other format PROJ
understands.}

\item{ctx}{A \code{\link[=proj_context]{proj_context()}}}

\item{source_crs, target_crs}{Source and/or target CRS definitions,
coerced using \code{\link[=as_proj]{as_proj()}}.}

\item{area}{A \link[=as_proj_area]{PROJ area} to use when selecting the
best available transformation.}

\item{options}{Options for instantiating the object by name}

\item{wkt}{A well-known text definition of the PROJ object.}

\item{pj}{A PROJ object or definition coerced using \code{\link[=as_proj]{as_proj()}}.}

\item{auth_name}{An authority name (e.g., "EPSG" or "OGC")}

\item{x}{An object to convert to a PROJ pointer}

\item{...}{Passed to S3 methods}
}
\description{
In PROJ speak, a "proj" can be a transformation, conversion,
CRS, ellipsoid, datum, or one of a few other types. The
CRS and conversion types are the most common and the
most useful but the other types are also occasionally
returned.
}
