% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_info}
\alias{proj_info}
\alias{proj_get_type}
\alias{proj_is_deprecated}
\alias{proj_is_equivalent_to}
\alias{proj_is_crs}
\alias{proj_get_remarks}
\alias{proj_get_scope}
\alias{proj_get_area_of_use}
\alias{proj_as_wkt}
\alias{proj_as_proj_string}
\alias{proj_as_projjson}
\title{PROJ object information}
\usage{
proj_info(obj)

proj_get_type(obj)

proj_is_deprecated(obj)

proj_is_equivalent_to(obj, other, criterion = NULL, ctx = proj_context())

proj_is_crs(obj)

proj_get_remarks(obj)

proj_get_scope(obj)

proj_get_area_of_use(obj)

proj_as_wkt(obj, wkt_type = NULL, options = NULL, ctx = proj_context())

proj_as_proj_string(
  obj,
  proj_string_type = NULL,
  options = NULL,
  ctx = proj_context()
)

proj_as_projjson(obj, options = NULL, ctx = proj_context())
}
\arguments{
\item{obj}{A \link[=proj_create]{PROJ object}}

\item{other}{Another \link[=proj_create]{PROJ object} against which
\code{obj} should be compared}

\item{criterion}{An equivalence criterion. One of "strict",
"equivalent", or "equivalent_except_axis_order_geogcrs".}

\item{ctx}{A \code{\link[=proj_context]{proj_context()}}}

\item{wkt_type}{One of "wkt2_2015", "wkt_2015_simplified",
"wkt2_2019", "wkt2_2019_simplified", "wkt1_gdal", or
"wkt1_esri". Use \code{NULL} for the recommended default
(wkt2_2019).}

\item{options}{String of additional options for various methods.}

\item{proj_string_type}{One of "proj_4" or "proj_5"}
}
\description{
PROJ object information
}
