% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{proj_trans}
\alias{proj_trans}
\alias{proj_trans.matrix}
\title{Transform coordinates}
\usage{
proj_trans(x, pj, direction = "fwd")

\method{proj_trans}{matrix}(x, pj, direction = "fwd")
}
\arguments{
\item{x}{An R object}

\item{pj}{A PROJ object or definition coerced using \code{\link[=as_proj]{as_proj()}}.}

\item{direction}{One of "fwd", "ident", or "inv".}
}
\value{
A transformed value of \code{x}
}
\description{
Transform coordinates
}
