% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlibproj-package.R
\docType{package}
\name{rlibproj-package}
\alias{rlibproj}
\alias{rlibproj-package}
\title{rlibproj: Generic Coordinate Transformation Library ('PROJ') R API}
\description{
Provides R bindings for the 'PROJ' coordinate transformation
  library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/rlibproj}
  \item Report bugs at \url{https://github.com/paleolimbot/rlibproj/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
